HTML-739: Refactor FormSubmissionActions#transitionToState(ProgramWorkflowState) into a utility method
